Homework 4 - Version 2
Cleaned up code a little bit. Implemented a method to check if the sides
form a triangle, as well as implementing a method to return the triangle
type if the sides are valid. ***WARNING ABOUT MY TESTS.  My tests passes
values into the new method I made and not the GetTriangle method***.